Remove total and stats from the public API.
